75: Research and refactor codebase
Reorganized folder structure
Created new route when updating a guide
