+++ src/components/pages/auth/Register.js
@@ -6,7 +6,7 @@ import { useDispatch } from "react-redux";
 
 import Form from "./common/Form";
 import { registerUser } from "../../../store/controllers/authController";
-import Alert from "../../common/Alert";
+import Alert from "../../ui/Alert";
 import messages from "../../../store/messages";
 import InputGroup from "./common/InputGroup";
