+++ src/components/pages/guides/FeaturedGuides.js
@@ -5,7 +5,7 @@ import { getGuidesByUserId } from "../../../store/controllers/guideController";
 import cardImg from '../../../assets/card_item.png';
 import { NavLink } from "react-router-dom";
 import { guideActions } from "../../../store/slices/guideSlice";
-import Loading from "../../common/Loading";
+import Loading from "../../ui/Loading";
 
 const FeaturedGuides = ({ userId }) => {
 	const dispatch = useDispatch();
@@ -21,23 +21,23 @@ const FeaturedGuides = ({ userId }) => {
 		<div className="w-[40%] bg-light-main p-5 rounded">
 			<h2 className=" text-2xl my-5">More from the author</h2>
 			{isLoading && <Loading />}
-				{guides && guides.map(guide =>
-					<NavLink to={`/guides/${guide.guideId}`} key={guide.guideId} className="flex mb-5 h-fit">
-						<img src={cardImg} alt="Preview" className="w-[30%] mr-5 rounded" />
-						<div>
-							<h1 className="text-secondary-main font-bold underline">{guide.title}</h1>
-							<p className=" font-light text-dark-main">
-								Creation date:{" "}
-								{new Date(guide.lastModified).getDate()}
-								-
-								{new Date(guide.lastModified).getMonth() + 1}
-								-
-								{new Date(guide.lastModified).getFullYear()}
+			{guides && guides.map(guide =>
+				<NavLink to={`/guides/${guide.guideId}`} key={guide.guideId} className="flex mb-5 h-fit">
+					<img src={cardImg} alt="Preview" className="w-[30%] mr-5 rounded" />
+					<div>
+						<h1 className="text-secondary-main font-bold underline">{guide.title}</h1>
+						<p className=" font-light text-dark-main">
+							Creation date:{" "}
+							{new Date(guide.lastModified).getDate()}
+							-
+							{new Date(guide.lastModified).getMonth() + 1}
+							-
+							{new Date(guide.lastModified).getFullYear()}
 
-							</p>
-						</div>
-					</NavLink>
-				)}
+						</p>
+					</div>
+				</NavLink>
+			)}
 		</div>
 	)
 }
